Rename test call scripts, update docs
It also adds AWS SDK S3 dependency
