+++ tools/psoxy-test/package.json
@@ -1,10 +1,10 @@
 {
   "name": "psoxy-test",
-  "version": "0.4.7",
+  "version": "0.4.8",
   "private": "true",
   "type": "module",
-  "description": "Worklytics Psoxy testing tool",
-  "main": "index.js",
+  "description": "Worklytics Psoxy testing tools",
+  "main": "psoxy-test-call.js",
   "scripts": {
     "test": "ava"
   },
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/Worklytics/psoxy#readme",
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.199.0",
     "aws4": "^1.11.0",
     "chalk": "^5.0.1",
     "commander": "^9.4.0",
